Some code updated
