+++ fronted/package.json
@@ -30,6 +30,7 @@
     "react-icons": "^5.3.0",
     "react-lottie-player": "^2.1.0",
     "react-paginate": "^8.2.0",
+    "react-player": "^2.16.0",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.22.1",
     "react-select": "^5.8.1",
